finishup intial availability
